chore: handle number values
